Add confirmation before deleting

+++ client/src/app/response-table/components/response-table.component.ts
@@ -32,6 +32,13 @@ export class ResponseTable {
   }
 
   handleDeletion(responseElem: ResponseTableElement) {
+    const message = 'Are you sure you want to delete this response? Doing so ' +
+                    'will remove all data related to this response including ' +
+                    'existing tags';
+    if(!confirm(message)) {
+      return;
+    }
+
     this.responseService.delete(responseElem.response);
 
     this.responseData = this.responseData.filter(elem => {
